requirements.txt
pandas
PyYAML
tqdm
langchain
langchain-community
langchain-openai
torch
transformers
datasets
accelerate
vllm
requests
fastchat
parmap
sentence_transformers
pyparsing
langchain_anthropic
plotly
kaleido
timeout_decorator
antlr4-python3-runtime==4.11.1
bitsandbytes